+++ utils/stake_test.go
@@ -302,14 +302,16 @@ func TestGetStake(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			retryMock := new(mocks.RetryUtils)
 			utilsMock := new(mocks.Utils)
+			stakeManagerMock := new(mocks.StakeManagerUtils)
 
 			optionsPackageStruct := OptionsPackageStruct{
-				RetryInterface: retryMock,
-				UtilsInterface: utilsMock,
+				RetryInterface:        retryMock,
+				UtilsInterface:        utilsMock,
+				StakeManagerInterface: stakeManagerMock,
 			}
 			utils := StartRazor(optionsPackageStruct)
 
-			utilsMock.On("GetStaker", mock.AnythingOfType("*ethclient.Client"), mock.AnythingOfType("uint32")).Return(tt.args.staker, tt.args.stakerErr)
+			stakeManagerMock.On("GetStaker", mock.AnythingOfType("*ethclient.Client"), mock.AnythingOfType("uint32")).Return(tt.args.staker, tt.args.stakerErr)
 			retryMock.On("RetryAttempts", mock.AnythingOfType("uint")).Return(retry.Attempts(1))
 
 			got, err := utils.GetStake(client, stakerId)
@@ -521,68 +523,6 @@ func TestGetStakeManagerWithOpts(t *testing.T) {
 	}
 	if !reflect.DeepEqual(gotStakeManager, stakeManager) {
 		t.Errorf("GetStakeManagerWithOpts() got stakeManager = %v, want %v", gotStakeManager, stakeManager)
-	}
-}
-
-func TestGetStakerSRZRBalance(t *testing.T) {
-	var (
-		client      *ethclient.Client
-		staker      bindings.StructsStaker
-		callOpts    bind.CallOpts
-		stakedToken *bindings.StakedToken
-	)
-
-	type args struct {
-		sRZR    *big.Int
-		sRZRErr error
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *big.Int
-		wantErr bool
-	}{
-		{
-			name: "Test 1: When GetStakerSRZRBalance executes successfully",
-			args: args{
-				sRZR:    big.NewInt(2000),
-				sRZRErr: nil,
-			},
-			want:    big.NewInt(2000),
-			wantErr: false,
-		},
-		{
-			name: "Test 2: When there is an error from BalanceOf()",
-			args: args{
-				sRZRErr: errors.New("sRZR error"),
-			},
-			want:    nil,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			utilsMock := new(mocks.Utils)
-			stakedTokenMock := new(mocks.StakedTokenUtils)
-
-			utilsMock.On("GetStakedToken", mock.Anything, mock.Anything).Return(stakedToken)
-			utilsMock.On("GetOptions").Return(callOpts)
-			stakedTokenMock.On("BalanceOf", mock.Anything, mock.Anything, mock.Anything).Return(tt.args.sRZR, tt.args.sRZRErr)
-
-			utils := StartRazor(OptionsPackageStruct{
-				UtilsInterface:       utilsMock,
-				StakedTokenInterface: stakedTokenMock,
-			})
-
-			got, err := utils.GetStakerSRZRBalance(client, staker)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("GetStakerSRZRBalance() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("GetStakerSRZRBalance() got = %v, want %v", got, tt.want)
-			}
-		})
 	}
 }
